+++ CHANGES.rst
@@ -1,6 +1,21 @@
 Changelog
 =========
 
+0.2 (**tip**)
+-------------
+
+Additions
+*********
+
+- Added context manager return to lock to allow use of the 'with'
+  statement.
+
+Features
+********
+
+- Configuration object to load/store configuration data in Zookeeper.
+
+
 0.1 (11/22/2011)
 ----------------